+++ src/components/chip/chip.tsx
@@ -5,27 +5,30 @@ import {
   Event,
   EventEmitter,
   h,
+  Host,
+  Listen,
   Method,
   Prop,
   State,
   VNode,
   Watch
 } from "@stencil/core";
+import { toAriaBoolean } from "../../utils/dom";
+import { CSS, SLOTS, ICONS } from "./resources";
+import { Appearance, Kind, Scale, SelectionMode } from "../interfaces";
 import {
   ConditionalSlotComponent,
   connectConditionalSlotComponent,
   disconnectConditionalSlotComponent
 } from "../../utils/conditionalSlot";
 import { slotChangeHasAssignedElement } from "../../utils/dom";
-import { guid } from "../../utils/guid";
 import {
   componentLoaded,
   LoadableComponent,
   setComponentLoaded,
   setUpLoadableComponent
 } from "../../utils/loadable";
-import { connectLocalized, disconnectLocalized, LocalizedComponent } from "../../utils/locale";
-import { createObserver } from "../../utils/observers";
+
 import {
   connectMessages,
   disconnectMessages,
@@ -33,9 +36,11 @@ import {
   T9nComponent,
   updateMessages
 } from "../../utils/t9n";
-import { Appearance, Kind, Scale } from "../interfaces";
+import { InteractiveComponent, updateHostInteraction } from "../../utils/interactive";
+import { connectLocalized, disconnectLocalized, LocalizedComponent } from "../../utils/locale";
+import { createObserver } from "../../utils/observers";
+import { isActivationKey } from "../../utils/key";
 import { ChipMessages } from "./assets/chip/t9n";
-import { CSS, ICONS, SLOTS } from "./resources";
 
 /**
  * @slot - A slot for adding text.
@@ -48,13 +53,28 @@ import { CSS, ICONS, SLOTS } from "./resources";
   assetsDirs: ["assets"]
 })
 export class Chip
-  implements ConditionalSlotComponent, LoadableComponent, LocalizedComponent, T9nComponent
+  implements
+    ConditionalSlotComponent,
+    InteractiveComponent,
+    LoadableComponent,
+    LocalizedComponent,
+    T9nComponent
 {
+  //--------------------------------------------------------------------------
+  //
+  //  Element
+  //
+  //--------------------------------------------------------------------------
+
+  @Element() el: HTMLCalciteChipElement;
+
   //--------------------------------------------------------------------------
   //
   //  Public Properties
   //
   //--------------------------------------------------------------------------
+  /** When true, interaction is prevented and the component is displayed with lower opacity. */
+  @Prop({ reflect: true }) disabled = false;
 
   /** Specifies the appearance style of the component. */
   @Prop({ reflect: true }) appearance: Extract<"outline" | "outline-fill" | "solid", Appearance> =
@@ -75,11 +95,26 @@ export class Chip
   /** Specifies the size of the component. */
   @Prop({ reflect: true }) scale: Scale = "m";
 
+  /** Accessible name for the component. */
+  @Prop() label: string;
+
   /** The component's value. */
   @Prop() value!: any;
 
   /** When `true`, hides the component. */
   @Prop({ reflect: true, mutable: true }) closed = false;
+
+  /**
+   * This internal property, managed by a containing `calcite-chip-group`, is
+   * conditionally set based on the `selectionMode` of the parent
+   *
+   * @internal
+   */
+  @Prop() selectionMode: Extract<"multiple" | "single" | "single-persist" | "none", SelectionMode> =
+    "none";
+
+  /** When true, the component is selected.  */
+  @Prop({ reflect: true, mutable: true }) selected = false;
 
   /**
    * Use this property to override individual strings used by the component.
@@ -100,13 +135,19 @@ export class Chip
     /* wired up by t9n util */
   }
 
-  // --------------------------------------------------------------------------
-  //
-  //  Private State/Properties
-  //
-  // --------------------------------------------------------------------------
+  /**
+   * When true, enables the chip to be focused, and allows the `calciteChipSelect` to emit.
+   * This is set to `true` by a parent Chip Group component.
+   *
+   * @internal
+   */
+  @Prop() interactive = false;
 
-  @Element() el: HTMLCalciteChipElement;
+  //--------------------------------------------------------------------------
+  //
+  //  Private State/Props
+  //
+  // --------------------------------------------------------------------------
 
   @State() defaultMessages: ChipMessages;
 
@@ -117,6 +158,39 @@ export class Chip
     updateMessages(this, this.effectiveLocale);
   }
 
+  @State() private hasText = false;
+
+  @State() private hasImage = false;
+
+  private containerEl: HTMLDivElement;
+
+  private closeButtonEl: HTMLButtonElement;
+
+  private parentGroupEl: HTMLCalciteChipGroupElement;
+
+  private mutationObserver = createObserver("mutation", () => this.updateHasText());
+
+  // --------------------------------------------------------------------------
+  //
+  //  Events
+  //
+  // --------------------------------------------------------------------------
+
+  /**
+   * Fires when the component's close button is selected.
+   */
+  @Event({ cancelable: false }) calciteChipClose: EventEmitter<void>;
+
+  /**
+   * Fires when the selected state of the component changes.
+   */
+  @Event({ cancelable: false }) calciteChipSelect: EventEmitter<void>;
+
+  /**
+   * @internal
+   */
+  @Event({ cancelable: false }) calciteInternalChipKeyEvent: EventEmitter<KeyboardEvent>;
+
   // --------------------------------------------------------------------------
   //
   //  Lifecycle
@@ -124,6 +198,7 @@ export class Chip
   // --------------------------------------------------------------------------
 
   connectedCallback(): void {
+    this.parentGroupEl = this.el.parentElement as HTMLCalciteChipGroupElement;
     connectConditionalSlotComponent(this);
     connectLocalized(this);
     connectMessages(this);
@@ -132,6 +207,10 @@ export class Chip
 
   componentDidLoad(): void {
     setComponentLoaded(this);
+  }
+
+  componentDidRender(): void {
+    updateHostInteraction(this);
   }
 
   disconnectedCallback(): void {
@@ -144,33 +223,58 @@ export class Chip
     setUpLoadableComponent(this);
     if (Build.isBrowser) {
       await setUpMessages(this);
-      this.updateHasContent();
+      this.updateHasText();
     }
   }
+  //--------------------------------------------------------------------------
+  //
+  //  Event Listeners
+  //
+  //--------------------------------------------------------------------------
+
+  @Listen("keydown")
+  keyDownHandler(event: KeyboardEvent): void {
+    if (event.target === this.el) {
+      switch (event.key) {
+        case " ":
+        case "Enter":
+          this.handleEmittingEvent();
+          event.preventDefault();
+          break;
+        case "ArrowRight":
+        case "ArrowLeft":
+        case "Home":
+        case "End":
+          this.calciteInternalChipKeyEvent.emit(event);
+          event.preventDefault();
+          break;
+      }
+    }
+  }
+
+  @Listen("click")
+  clickHandler(): void {
+    if (!this.interactive && this.closable) {
+      this.closeButtonEl.focus();
+    }
+  }
+
   //--------------------------------------------------------------------------
   //
   //  Public Methods
   //
   //--------------------------------------------------------------------------
 
-  /** When `closable` is `true`, sets focus on the component's "close" button (the first focusable item). */
+  /** Sets focus on the component. */
   @Method()
   async setFocus(): Promise<void> {
     await componentLoaded(this);
-
-    this.closeButton?.focus();
+    if (!this.disabled && this.interactive) {
+      this.containerEl?.focus();
+    } else if (!this.disabled && this.closable) {
+      this.closeButtonEl?.focus();
+    }
   }
-
-  // --------------------------------------------------------------------------
-  //
-  //  Events
-  //
-  // --------------------------------------------------------------------------
-
-  /**
-   * Fires when the close button is clicked.
-   */
-  @Event({ cancelable: false }) calciteChipClose: EventEmitter<void>;
 
   // --------------------------------------------------------------------------
   //
@@ -178,18 +282,21 @@ export class Chip
   //
   // --------------------------------------------------------------------------
 
-  closeClickHandler = (event: MouseEvent): void => {
-    event.preventDefault();
+  private close = (): void => {
     this.calciteChipClose.emit();
+    this.selected = false;
     this.closed = true;
   };
 
-  private updateHasContent() {
-    const slottedContent = this.el.textContent.trim().length > 0 || this.el.childNodes.length > 0;
-    this.hasContent =
-      this.el.childNodes.length > 0 && this.el.childNodes[0]?.nodeName === "#text"
-        ? this.el.textContent?.trim().length > 0
-        : slottedContent;
+  private closeButtonKeyDownHandler = (event: KeyboardEvent): void => {
+    if (isActivationKey(event.key)) {
+      event.preventDefault();
+      this.close();
+    }
+  };
+
+  private updateHasText() {
+    this.hasText = this.el.textContent.trim().length > 0;
   }
 
   private setupTextContentObserver() {
@@ -200,24 +307,12 @@ export class Chip
     this.hasImage = slotChangeHasAssignedElement(event);
   };
 
-  //--------------------------------------------------------------------------
-  //
-  //  Private State/Props
-  //
-  //--------------------------------------------------------------------------
+  private handleEmittingEvent = (): void => {
+    if (this.interactive) {
+      this.calciteChipSelect.emit();
+    }
+  };
 
-  /** watches for changing text content */
-  private mutationObserver = createObserver("mutation", () => this.updateHasContent());
-
-  private closeButton: HTMLButtonElement;
-
-  private guid: string = guid();
-
-  /** determine if there is slotted content for styling purposes */
-  @State() private hasContent = false;
-
-  /** determine if there is slotted image for styling purposes */
-  @State() private hasImage = false;
   //--------------------------------------------------------------------------
   //
   //  Render Methods
@@ -226,14 +321,52 @@ export class Chip
 
   renderChipImage(): VNode {
     return (
-      <div class={CSS.imageContainer} key="image">
+      <div class={CSS.imageContainer}>
         <slot name={SLOTS.image} onSlotchange={this.handleSlotImageChange} />
       </div>
     );
   }
 
-  render(): VNode {
-    const iconEl = (
+  renderSelectionIcon(): VNode {
+    const icon =
+      this.selectionMode === "multiple" && this.selected
+        ? ICONS.checked
+        : this.selectionMode === "multiple"
+        ? ICONS.unchecked
+        : this.selected
+        ? ICONS.checkedSingle
+        : undefined;
+
+    return (
+      <div
+        class={{
+          [CSS.selectIcon]: true,
+          [CSS.selectIconActive]: this.selectionMode === "multiple" || this.selected
+        }}
+      >
+        <calcite-icon icon={icon} scale={this.scale === "l" ? "m" : "s"} />
+      </div>
+    );
+  }
+
+  renderCloseButton(): VNode {
+    return (
+      <button
+        aria-label={this.messages.dismissLabel}
+        class={CSS.close}
+        onClick={this.close}
+        onKeyDown={this.closeButtonKeyDownHandler}
+        tabIndex={this.disabled ? -1 : 0}
+        // eslint-disable-next-line react/jsx-sort-props
+        ref={(el) => (this.closeButtonEl = el)}
+      >
+        <calcite-icon icon={ICONS.close} scale={this.scale === "l" ? "m" : "s"} />
+      </button>
+    );
+  }
+
+  renderIcon(): VNode {
+    return (
       <calcite-icon
         class={CSS.chipIcon}
         flipRtl={this.iconFlipRtl}
@@ -241,39 +374,53 @@ export class Chip
         scale={this.scale === "l" ? "m" : "s"}
       />
     );
+  }
 
-    const closeButton = (
-      <button
-        aria-describedby={this.guid}
-        aria-label={this.messages.dismissLabel}
-        class={CSS.close}
-        onClick={this.closeClickHandler}
-        // eslint-disable-next-line react/jsx-sort-props
-        ref={(el) => (this.closeButton = el)}
-      >
-        <calcite-icon
-          class={CSS.closeIcon}
-          icon={ICONS.close}
-          scale={this.scale === "l" ? "m" : "s"}
-        />
-      </button>
-    );
-
+  render(): VNode {
+    const disableInteraction = this.disabled || (!this.disabled && !this.interactive);
     return (
-      <div
-        class={{
-          [CSS.container]: true,
-          [CSS.contentSlotted]: this.hasContent,
-          [CSS.imageSlotted]: this.hasImage
-        }}
-      >
-        {this.renderChipImage()}
-        {this.icon ? iconEl : null}
-        <span class={CSS.title} id={this.guid}>
-          <slot />
-        </span>
-        {this.closable ? closeButton : null}
-      </div>
+      <Host>
+        <div
+          aria-checked={this.interactive ? toAriaBoolean(this.selected) : undefined}
+          aria-disabled={disableInteraction ? toAriaBoolean(this.disabled) : undefined}
+          aria-label={this.label}
+          aria-labelledby={this.parentGroupEl.label}
+          class={{
+            [CSS.container]: true,
+            [CSS.textSlotted]: this.hasText,
+            [CSS.imageSlotted]: this.hasImage,
+            [CSS.selectable]: this.selectionMode !== "none",
+            [CSS.multiple]: this.selectionMode === "multiple",
+            [CSS.closable]: this.closable,
+            [CSS.nonInteractive]: !this.interactive,
+            [CSS.isCircle]:
+              !this.closable &&
+              !this.hasText &&
+              (!this.icon || !this.hasImage) &&
+              (this.selectionMode === "none" ||
+                (!!this.selectionMode && this.selectionMode !== "multiple" && !this.selected))
+          }}
+          onClick={this.handleEmittingEvent}
+          role={
+            this.selectionMode === "multiple" && this.interactive
+              ? "checkbox"
+              : this.interactive
+              ? "radio"
+              : undefined
+          }
+          tabIndex={disableInteraction ? -1 : 0}
+          // eslint-disable-next-line react/jsx-sort-props
+          ref={(el) => (this.containerEl = el)}
+        >
+          {this.selectionMode !== "none" && this.renderSelectionIcon()}
+          {this.renderChipImage()}
+          {this.icon && this.renderIcon()}
+          <span class={CSS.title}>
+            <slot />
+          </span>
+          {this.closable && this.renderCloseButton()}
+        </div>
+      </Host>
     );
   }
 }