+++ src/components/date-picker-day/date-picker-day.tsx
@@ -167,7 +167,7 @@ export class DatePickerDay implements InteractiveComponent {
         onKeyDown={this.keyDownHandler}
         role="button"
       >
-        <div class={{ "day-v-wrapper": true, [CSS_UTILITY.rtl]: dir === "rtl" }}>
+        <div aria-hidden="true" class={{ "day-v-wrapper": true, [CSS_UTILITY.rtl]: dir === "rtl" }}>
           <div class="day-wrapper">
             <span class="day">
               <span class="text">{formattedDay}</span>