+++ src/components/input-date-picker/input-date-picker.tsx
@@ -68,6 +68,7 @@ import { HeadingLevel } from "../functional/Heading";
 import { CSS } from "./resources";
 import { connectMessages, disconnectMessages, setUpMessages, T9nComponent } from "../../utils/t9n";
 import { InputDatePickerMessages } from "./assets/input-date-picker/t9n";
+import { guid } from "../../utils/guid";
 
 @Component({
   tag: "calcite-input-date-picker",
@@ -483,15 +484,20 @@ export class InputDatePicker
     };
 
     return (
-      <Host onBlur={this.deactivate} onKeyDown={this.keyDownHandler} role="application">
+      <Host onBlur={this.deactivate} onKeyDown={this.keyDownHandler}>
         {this.localeData && (
-          <div aria-expanded={toAriaBoolean(this.open)} class="input-container" role="application">
+          <div class="input-container">
             <div
               class="input-wrapper"
+              onClick={this.onInputWrapperClick}
               // eslint-disable-next-line react/jsx-sort-props
               ref={this.setStartWrapper}
             >
               <calcite-input
+                aria-autocomplete="none"
+                aria-controls={this.dialogId}
+                aria-expanded={toAriaBoolean(this.open)}
+                aria-haspopup="dialog"
                 class={`input ${
                   this.layout === "vertical" && this.range ? `no-bottom-border` : ``
                 }`}
@@ -503,22 +509,27 @@ export class InputDatePicker
                 onCalciteInputInput={this.calciteInternalInputInputHandler}
                 onCalciteInternalInputBlur={this.calciteInternalInputBlurHandler}
                 onCalciteInternalInputFocus={this.startInputFocus}
+                onFocus={this.startEndInputFocus}
                 placeholder={this.localeData?.placeholder}
                 readOnly={readOnly}
+                role="combobox"
                 scale={this.scale}
                 type="text"
                 // eslint-disable-next-line react/jsx-sort-props
                 ref={this.setStartInput}
               />
+              {this.renderToggleIcon(this.open && this.focusedInput === "start")}
             </div>
             <div
               aria-hidden={toAriaBoolean(!this.open)}
               aria-label={messages.chooseDate}
+              aria-live="polite"
               aria-modal="true"
               class={{
                 [CSS.menu]: true,
                 [CSS.menuActive]: this.open
               }}
+              id={this.dialogId}
               role="dialog"
               // eslint-disable-next-line react/jsx-sort-props
               ref={this.setFloatingEl}
@@ -571,10 +582,15 @@ export class InputDatePicker
             {this.range && (
               <div
                 class="input-wrapper"
+                onClick={this.onInputWrapperClick}
                 // eslint-disable-next-line react/jsx-sort-props
                 ref={this.setEndWrapper}
               >
                 <calcite-input
+                  aria-autocomplete="none"
+                  aria-controls={this.dialogId}
+                  aria-expanded={toAriaBoolean(this.open)}
+                  aria-haspopup="dialog"
                   class={{
                     input: true,
                     "border-top-color-one": this.layout === "vertical" && this.range
@@ -586,13 +602,16 @@ export class InputDatePicker
                   onCalciteInputInput={this.calciteInternalInputInputHandler}
                   onCalciteInternalInputBlur={this.calciteInternalInputBlurHandler}
                   onCalciteInternalInputFocus={this.endInputFocus}
+                  onFocus={this.startEndInputFocus}
                   placeholder={this.localeData?.placeholder}
                   readOnly={readOnly}
+                  role="combobox"
                   scale={this.scale}
                   type="text"
                   // eslint-disable-next-line react/jsx-sort-props
                   ref={this.setEndInput}
                 />
+                {this.renderToggleIcon(this.open && this.focusedInput === "end")}
               </div>
             )}
           </div>
@@ -602,11 +621,21 @@ export class InputDatePicker
     );
   }
 
+  renderToggleIcon(open: boolean): VNode {
+    return (
+      <span class={CSS.toggleIcon}>
+        <calcite-icon icon={open ? "chevron-up" : "chevron-down"} scale="s" />
+      </span>
+    );
+  }
+
   //--------------------------------------------------------------------------
   //
   //  Private State/Props
   //
   //--------------------------------------------------------------------------
+
+  private dialogId = `date-picker-dialog--${guid()}`;
 
   filteredFlipPlacements: EffectivePlacement[];
 
@@ -623,6 +652,8 @@ export class InputDatePicker
   @State() effectiveLocale = "";
 
   @State() focusedInput: "start" | "end" = "start";
+
+  private lastBlurredInput: "start" | "end" | "none" = "none";
 
   @State() private localeData: DateLocaleData;
 
@@ -664,6 +695,16 @@ export class InputDatePicker
   //  Private Methods
   //
   //--------------------------------------------------------------------------
+
+  private onInputWrapperClick = () => {
+    if (this.range && this.lastBlurredInput !== "none" && this.open) {
+      // we keep the date-picker open when moving between inputs
+    } else {
+      this.open = !this.open;
+    }
+
+    this.lastBlurredInput = "none";
+  };
 
   setFilteredPlacements = (): void => {
     const { el, flipPlacements } = this;
@@ -708,6 +749,7 @@ export class InputDatePicker
 
   deactivate = (): void => {
     this.open = false;
+    this.lastBlurredInput = "none";
   };
 
   private commitValue(): void {
@@ -755,7 +797,12 @@ export class InputDatePicker
 
   keyDownHandler = (event: KeyboardEvent): void => {
     const { defaultPrevented, key } = event;
-    if (key === "Enter" && !defaultPrevented) {
+
+    if (defaultPrevented) {
+      return;
+    }
+
+    if (key === "Enter") {
       this.commitValue();
       if (this.shouldFocusRangeEnd()) {
         this.endInput?.setFocus();
@@ -765,23 +812,26 @@ export class InputDatePicker
       if (submitForm(this)) {
         event.preventDefault();
       }
-    } else if (key === "Escape" && !defaultPrevented) {
+    } else if (key === "ArrowDown") {
+      this.open = true;
+      event.preventDefault();
+    } else if (key === "Escape") {
       this.open = false;
       event.preventDefault();
     }
   };
 
   startInputFocus = (): void => {
-    if (!this.readOnly) {
-      this.open = true;
-    }
     this.focusedInput = "start";
   };
 
+  startEndInputFocus = (event: FocusEvent): void => {
+    const blurredEl = event.relatedTarget as HTMLElement;
+    this.lastBlurredInput =
+      blurredEl === this.startInput ? "start" : blurredEl === this.endInput ? "end" : "none";
+  };
+
   endInputFocus = (): void => {
-    if (!this.readOnly) {
-      this.open = true;
-    }
     this.focusedInput = "end";
   };